requirements.txt
torch>=1.7.0
torchvision
detectron2
git+https://github.com/facebookresearch/detectron2.git
numpy
matplotlib
Pillow
PyYAML
opencv-python
pandas
